requirements.txt
alembic==1.13.1
bcrypt==4.1.3
cachetools==5.3.3
fastapi==0.111.0
mysqlclient==2.2.4
pydantic==2.7.3
python-dotenv==1.0.1
python-jose==3.3.0
SQLAlchemy==2.0.30
uvicorn==0.30.1

